FEAT: 리뷰 api 수정

+++ src/main/java/com/goomoong/room9backend/service/ReviewService.java
@@ -4,6 +4,7 @@ import com.goomoong.room9backend.domain.review.Review;
 import com.goomoong.room9backend.domain.review.dto.ReviewSearchDto;
 import com.goomoong.room9backend.domain.review.dto.scoreDto;
 import com.goomoong.room9backend.exception.NoSuchReviewException;
+import com.goomoong.room9backend.exception.NotAllowedUserException;
 import com.goomoong.room9backend.repository.review.ReviewRepository;
 import com.goomoong.room9backend.util.AboutScore;
 import lombok.RequiredArgsConstructor;
@@ -27,14 +28,22 @@ public class ReviewService {
     }
 
     @Transactional
-    public void update(Long id,  String reviewContent, int reviewScore){
+    public void update(Long userId, Long id,  String reviewContent, int reviewScore){
         Review review =  reviewRepository.findById(id).orElseThrow(() -> new NoSuchReviewException("존재하지 않는 리뷰입니다."));
+
+        if (userId != review.getUser().getId())
+            throw new NotAllowedUserException("권한이 없는 사용자입니다.");
 
         review.update(reviewContent, reviewScore);
     }
 
     @Transactional
-    public void delete(Long id){
+    public void delete(Long userId, Long id){
+        Review review = reviewRepository.findById(id).orElseThrow(() -> new NoSuchReviewException("존재하지 않는 리뷰입니다."));
+
+        if (userId != review.getUser().getId())
+            throw new NotAllowedUserException("권한이 없는 사용자입니다.");
+
         reviewRepository.delete(findById(id));
     }
